cdk/bidiagent/requirements.txt
bedrock-agentcore
strands-agents-tools
strands-agents[bidi]
pyaudio>=0.2.14
starlette
